oopo/requirements.txt
py-cord
pymongo
emoji
python-dotenv
